feat: implemented app features > Implemented search of arrival time when user touch on bus stop > Impelmented pop-up with line info when user touch on bus time while serch bus stop > Implemented automatic authenticate when server return 401 or 403 and save cookie for next requests > Created new bus and bus stop icons to show on map > Implemented automatic refresh of bus positions on map

+++ app/src/main/java/com/matreis/teste/sptrans/domain/model/LineWithVehicles.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class LineWithVehicles(
     @SerializedName("c"   ) var fullSign   : String?       = null,
-    @SerializedName("cl"  ) var lineCode  : Int?          = null,
+    @SerializedName("cl"  ) var lineCode  : Long?          = null,
     @SerializedName("sl"  ) var direction  : Int?          = null,
     @SerializedName("lt0" ) var destinationSign : String?       = null,
     @SerializedName("lt1" ) var originSign : String?       = null,
